r/aws_glue_partition_index: Reduce visibility.
