+++ internal/service/glue/crawler.go
@@ -649,6 +649,30 @@ func resourceCrawlerDelete(ctx context.Context, d *schema.ResourceData, meta any
 	return diags
 }
 
+func findCrawlerByName(ctx context.Context, conn *glue.Client, name string) (*awstypes.Crawler, error) {
+	input := &glue.GetCrawlerInput{
+		Name: aws.String(name),
+	}
+
+	output, err := conn.GetCrawler(ctx, input)
+	if errs.IsA[*awstypes.EntityNotFoundException](err) {
+		return nil, &retry.NotFoundError{
+			LastError:   err,
+			LastRequest: input,
+		}
+	}
+
+	if err != nil {
+		return nil, err
+	}
+
+	if output == nil || output.Crawler == nil {
+		return nil, tfresource.NewEmptyResultError(input)
+	}
+
+	return output.Crawler, nil
+}
+
 func createCrawlerInput(ctx context.Context, d *schema.ResourceData, crawlerName string) (*glue.CreateCrawlerInput, error) {
 	crawlerInput := &glue.CreateCrawlerInput{
 		Name:         aws.String(crawlerName),
