+++ internal/service/glue/partition_index.go
@@ -12,9 +12,11 @@ import (
 	"github.com/aws/aws-sdk-go-v2/service/glue"
 	awstypes "github.com/aws/aws-sdk-go-v2/service/glue/types"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
+	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/retry"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/validation"
 	"github.com/hashicorp/terraform-provider-aws/internal/conns"
+	"github.com/hashicorp/terraform-provider-aws/internal/enum"
 	"github.com/hashicorp/terraform-provider-aws/internal/errs"
 	"github.com/hashicorp/terraform-provider-aws/internal/errs/sdkdiag"
 	"github.com/hashicorp/terraform-provider-aws/internal/flex"
@@ -23,11 +25,12 @@ import (
 )
 
 // @SDKResource("aws_glue_partition_index", name="Partition Index")
-func ResourcePartitionIndex() *schema.Resource {
+func resourcePartitionIndex() *schema.Resource {
 	return &schema.Resource{
 		CreateWithoutTimeout: resourcePartitionIndexCreate,
 		ReadWithoutTimeout:   resourcePartitionIndexRead,
 		DeleteWithoutTimeout: resourcePartitionIndexDelete,
+
 		Importer: &schema.ResourceImporter{
 			StateContext: schema.ImportStatePassthroughContext,
 		},
@@ -124,7 +127,7 @@ func resourcePartitionIndexRead(ctx context.Context, d *schema.ResourceData, met
 	}
 
 	log.Printf("[DEBUG] Reading Glue Partition Index: %s", d.Id())
-	partition, err := FindPartitionIndexByName(ctx, conn, d.Id())
+	partition, err := findPartitionIndexByName(ctx, conn, d.Id())
 	if !d.IsNewResource() && tfresource.NotFound(err) {
 		log.Printf("[WARN] Glue Partition Index (%s) not found, removing from state", d.Id())
 		d.SetId("")
@@ -174,6 +177,104 @@ func resourcePartitionIndexDelete(ctx context.Context, d *schema.ResourceData, m
 	}
 
 	return diags
+}
+
+func findPartitionIndexByName(ctx context.Context, conn *glue.Client, id string) (*awstypes.PartitionIndexDescriptor, error) {
+	catalogID, dbName, tableName, partIndex, err := readPartitionIndexID(id)
+	if err != nil {
+		return nil, err
+	}
+
+	input := &glue.GetPartitionIndexesInput{
+		CatalogId:    aws.String(catalogID),
+		DatabaseName: aws.String(dbName),
+		TableName:    aws.String(tableName),
+	}
+
+	var result *awstypes.PartitionIndexDescriptor
+
+	output, err := conn.GetPartitionIndexes(ctx, input)
+
+	if errs.IsA[*awstypes.EntityNotFoundException](err) {
+		return nil, &retry.NotFoundError{
+			LastError:   err,
+			LastRequest: input,
+		}
+	}
+
+	if err != nil {
+		return nil, err
+	}
+
+	if output == nil {
+		return nil, tfresource.NewEmptyResultError(input)
+	}
+
+	for _, partInd := range output.PartitionIndexDescriptorList {
+		if aws.ToString(partInd.IndexName) == partIndex {
+			result = &partInd
+			break
+		}
+	}
+
+	if result == nil {
+		return nil, &retry.NotFoundError{
+			LastError:   err,
+			LastRequest: input,
+		}
+	}
+
+	return result, nil
+}
+
+func statusPartitionIndex(ctx context.Context, conn *glue.Client, id string) retry.StateRefreshFunc {
+	return func() (any, string, error) {
+		output, err := findPartitionIndexByName(ctx, conn, id)
+
+		if tfresource.NotFound(err) {
+			return nil, "", nil
+		}
+
+		if err != nil {
+			return nil, "", err
+		}
+
+		return output, string(output.IndexStatus), nil
+	}
+}
+
+func waitPartitionIndexCreated(ctx context.Context, conn *glue.Client, id string, timeout time.Duration) (*awstypes.PartitionIndexDescriptor, error) {
+	stateConf := &retry.StateChangeConf{
+		Pending: enum.Slice(awstypes.PartitionIndexStatusCreating),
+		Target:  enum.Slice(awstypes.PartitionIndexStatusActive),
+		Refresh: statusPartitionIndex(ctx, conn, id),
+		Timeout: timeout,
+	}
+
+	outputRaw, err := stateConf.WaitForStateContext(ctx)
+
+	if output, ok := outputRaw.(*awstypes.PartitionIndexDescriptor); ok {
+		return output, err
+	}
+
+	return nil, err
+}
+
+func waitPartitionIndexDeleted(ctx context.Context, conn *glue.Client, id string, timeout time.Duration) (*awstypes.PartitionIndexDescriptor, error) {
+	stateConf := &retry.StateChangeConf{
+		Pending: enum.Slice(awstypes.PartitionIndexStatusDeleting),
+		Target:  []string{},
+		Refresh: statusPartitionIndex(ctx, conn, id),
+		Timeout: timeout,
+	}
+
+	outputRaw, err := stateConf.WaitForStateContext(ctx)
+
+	if output, ok := outputRaw.(*awstypes.PartitionIndexDescriptor); ok {
+		return output, err
+	}
+
+	return nil, err
 }
 
 func expandPartitionIndex(l []any) *awstypes.PartitionIndex {
